refactor: consolidate registry modules
