+++ src/entity/core/registry_validator.py
