--- src/registry/__init__.py
@@ -1,13 +1,0 @@
-from __future__ import annotations
-
-"""Registry module exports for easy access."""
-
-from .registries import PluginRegistry, SystemRegistries, ToolRegistry
-from .validator import RegistryValidator
-
-__all__ = [
-    "PluginRegistry",
-    "SystemRegistries",
-    "ToolRegistry",
-    "RegistryValidator",
-]
