+++ tests/integration/test_stage_failures.py
@@ -8,7 +8,7 @@ from pipeline.base_plugins import BasePlugin
 from pipeline.pipeline import generate_pipeline_id
 from pipeline.resources import ResourceContainer
 from pipeline.state import ConversationEntry, MetricsCollector, PipelineState
-from registry import PluginRegistry, SystemRegistries, ToolRegistry
+from entity.core.registries import PluginRegistry, SystemRegistries, ToolRegistry
 
 from entity.core.state_logger import LogReplayer, StateLogger
